# docs/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Advanced web panel for managing Xray-core servers. Supports VMESS, VLESS, Trojan, Shadowsocks, Hysteria, WireGuard, and more.",
        "title": "3X-UI Panel API",
        "termsOfService": "https://github.com/MHSanaei/3x-ui",
        "contact": {
            "name": "MHSanaei",
            "url": "https://github.com/MHSanaei/3x-ui"
        },
        "license": {
            "name": "GNU GPL v3",
            "url": "https://www.gnu.org/licenses/gpl-3.0.html"
        },
        "version": "3.0.1"
    },
    "host": "localhost:2053",
    "basePath": "/",
    "paths": {
...
        "/panel/api/custom-geo/add": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Registers a custom geo source. Alias is auto-normalised; URL must point to a .dat or .json file.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Custom Geo"
                ],
                "summary": "Add custom geo source",
                "parameters": [
                    {
                        "description": "Custom geo source details",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controller.customGeoForm"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entity.Msg"
                        }
                    }
                }
            }
        },
        "/panel/api/custom-geo/aliases": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns geo aliases usable in routing rules — both built-in defaults and user-configured ones.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Custom Geo"
                ],
                "summary": "List geo aliases",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entity.Msg"
                        }
                    }
                }
            }
        },
        "/panel/api/custom-geo/delete/{id}": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Removes a custom geo source and its cached file from disk.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Custom Geo"
                ],
                "summary": "Delete custom geo source",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Custom geo source ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entity.Msg"
                        }
                    }
                }
            }
        },
        "/panel/api/custom-geo/download/{id}": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Triggers a re-download of a custom geo source from its URL.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Custom Geo"
                ],
                "summary": "Download custom geo file",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Custom geo source ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entity.Msg"
                        }
                    }
                }
            }
        },
        "/panel/api/custom-geo/list": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns all configured custom geo sources with type, alias, URL, status, and download timestamps.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Custom Geo"
                ],
                "summary": "List custom geo sources",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entity.Msg"
                        }
                    }
                }
            }
        },
        "/panel/api/custom-geo/update-all": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Re-downloads every configured custom geo source. Errors are reported per-source in the response.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Custom Geo"
                ],
                "summary": "Update all custom geo sources",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entity.Msg"
                        }
                    }
                }
            }
        },
        "/panel/api/custom-geo/update/{id}": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Replaces a custom geo source configuration. Body shape matches /add.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Custom Geo"
                ],
                "summary": "Update custom geo source",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Custom geo source ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Updated source details",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controller.customGeoForm"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entity.Msg"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "controller.customGeoForm": {
            "type": "object",
            "properties": {
                "alias": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                }
            }
        },
...
        "entity.Msg": {
            "type": "object",
            "properties": {
                "msg": {
                    "description": "Response message text",
                    "type": "string"
                },
                "obj": {
                    "description": "Optional data object"
                },
                "success": {
                    "description": "Indicates if the operation was successful",
                    "type": "boolean"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Bearer token authentication. Use format: \"Bearer \u003ctoken\u003e\". Get your token from Settings → Security → API Token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}